requirements.txt
asgiref==3.8.1
Django==5.0.7
django-ckeditor==6.7.1
django-js-asset==2.2.0
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
